# swagger.yaml
openapi: 3.0.0
info:
  title: Quinielas API
  description: Swagger Quinielas - OpenAPI 3.0.
  version: 0.0.1
servers:
  - url: http://localhost:8080/api
    description: Server for testing
...
paths:
...
  /user:
    post:
      tags:
        - user
      summary: Adds a new user to the database
      description: New user will be added if not exist in database. Only supported for normal users, not admins
      operationId: AddUser
      requestBody:
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User schema from mongoose
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'                    
  /user/{userId}:
    delete:
      tags:
        - user
      security:
        - bearerAuth: []
      summary: Deletes a user from the database
      description: Delete a user from the database if it exists
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: user to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'User succesfully deleted from database'
        '400':
          description: 'Invalid id provided'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 
  schemas:
...
    User:
      required:
        - name
        - email
        - password
      type: object
      properties:
        name:
          type: string
          example: Ismael
        email:
          type: string
          example: icordondominguez@gmail.com
        password:
          type: string
          example: 7sept2024
...